+++ test/unit/utils/TimeRange.test.js
@@ -5,15 +5,23 @@ function time (y, m = 1, d = 1, h = 0, min = 0, sec = 0) {
 }
 
 function startTimeValue (iso) {
-    return TimeRange.fromText(iso).startValue;
+    const timeZone = null;
+    const start = null;
+    const end = null;
+    return new TimeRange(timeZone, iso, start, end).startValue;
 }
 
 function endTimeValue (iso) {
-    return TimeRange.fromText(iso).endValue;
+    const timeZone = null;
+    const start = null;
+    const end = null;
+    return new TimeRange(timeZone, iso, start, end).endValue;
 }
 
 function ISOperiod (start, end) {
-    return TimeRange.fromStartEndValues(start, end).text;
+    const timeZone = null;
+    const iso = null;
+    return new TimeRange(timeZone, iso, start, end).iso;
 }
 
 describe('TimeRange defined by a ISO string', () => {